+++ src/Widgets/oxidestablewidget.h
@@ -11,6 +11,8 @@
 
 #include <QWidget>
 
+#include "data.h"
+
 namespace Ui
 {
     class OxidesTableWidget;
@@ -22,9 +24,9 @@ class OxidesTableWidget : public QWidget
 
     public:
         explicit OxidesTableWidget(QWidget *parent = 0);
-        OxidesTableWidget(std::vector<double> contents, QWidget *parent = 0);
+        OxidesTableWidget(std::vector<Float> contents, QWidget *parent = 0);
         ~OxidesTableWidget();
-        std::vector<double> getOxidesContent();
+        std::vector<Float> getOxidesContent();
 
     private:
         void fillTable();
